requirements.txt
llama-parse
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
python-dotenv   
IPython